Update the stored procedure helpers in the TPCE benchmark
