+++ src/main/java/org/vanilladb/bench/benchmarks/tpce/rte/TpceSchemaBuilderParamHelper.java
@@ -1,26 +1,6 @@
-/*******************************************************************************
- * Copyright 2016, 2017 vanilladb.org contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *******************************************************************************/
-package org.vanilladb.bench.server.param.tpce;
+package org.vanilladb.bench.benchmarks.tpce.rte;
 
-import org.vanilladb.core.sql.Schema;
-import org.vanilladb.core.sql.storedprocedure.SpResultRecord;
-import org.vanilladb.core.sql.storedprocedure.StoredProcedureParamHelper;
-
-public class TpceSchemaBuilderParamHelper extends StoredProcedureParamHelper {
-
+public class TpceSchemaBuilderParamHelper {
 	private static final String TABLES_DDL[] = {
 			// Customer Category
 			// TODO: watch_item, watch_list
@@ -93,20 +73,5 @@ public class TpceSchemaBuilderParamHelper extends StoredProcedureParamHelper {
 
 	public String[] getIndexSchemas() {
 		return INDEXES_DDL;
-	}
-
-	@Override
-	public void prepareParameters(Object... pars) {
-		// nothing to do
-	}
-
-	@Override
-	public Schema getResultSetSchema() {
-		return new Schema();
-	}
-
-	@Override
-	public SpResultRecord newResultSetRecord() {
-		return new SpResultRecord();
 	}
 }
